+++ Editor/SpriteLib/SpriteResolverInspector.cs
@@ -66,6 +66,8 @@ namespace UnityEditor.U2D.Animation
         
         SpriteResolver spriteResolver => target as SpriteResolver;
 
+        bool IsSpriteHashAssigned => SpriteResolver.ConvertFloatToInt(m_SpriteKey.floatValue) != 0;
+
         void GetCategoryAndLabelStringValue(out string categoryName, out string labelName)
         {
             categoryName = null;
@@ -73,12 +75,12 @@ namespace UnityEditor.U2D.Animation
             var spriteLib = spriteResolver.spriteLibrary;
             if (spriteLib != null)
             {
-                int entryHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_SpriteKey.floatValue));
+                var entryHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_SpriteKey.floatValue));
                 spriteLib.GetCategoryAndEntryNameFromHash(entryHash, out categoryName, out labelName);
-                if (string.IsNullOrEmpty(categoryName) || string.IsNullOrEmpty(labelName))
+                if (!IsSpriteHashAssigned && (string.IsNullOrEmpty(categoryName) || string.IsNullOrEmpty(labelName)))
                 {
-                    int labelHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_LabelHash.floatValue));
-                    int categoryHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_CategoryHash.floatValue));
+                    var labelHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_LabelHash.floatValue));
+                    var categoryHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_CategoryHash.floatValue));
                     m_SpriteKey.floatValue = SpriteResolver.ConvertCategoryLabelHashToSpriteKey(spriteLib, categoryHash, labelHash);
                     entryHash = SpriteLibraryUtility.Convert32BitTo30BitHash(SpriteResolver.ConvertFloatToInt(m_SpriteKey.floatValue));
                     spriteLib.GetCategoryAndEntryNameFromHash(entryHash, out categoryName, out labelName);
@@ -172,9 +174,8 @@ namespace UnityEditor.U2D.Animation
                 UpdateSpriteLibrary();
 
             GetCategoryAndLabelStringValue(out var currentCategoryValue, out var currentLabelValue);
-            var catIndex = Array.FindIndex(m_CategorySelection, x => x == currentCategoryValue);
-            if (catIndex >= 0)
-                m_CategorySelectionIndex = catIndex;
+
+            m_CategorySelectionIndex = Array.FindIndex(m_CategorySelection, x => x == currentCategoryValue);
             ValidateCategorySelectionIndexValue();
             
             EditorGUI.BeginChangeCheck();
@@ -229,18 +230,13 @@ namespace UnityEditor.U2D.Animation
 
             if (m_PreviousCategoryValue != currentCategoryValue)
             {
-                if (!string.IsNullOrEmpty(currentCategoryValue))
-                {
-                    if (m_SpriteLibSelection.ContainsKey(currentCategoryValue))
-                    {
-                        m_SpriteSelectorWidget.UpdateContents(m_SpriteLibSelection[currentCategoryValue].sprites);
-                    }
-                    else
-                        m_SpriteSelectorWidget.UpdateContents(new Sprite[0]);
-                    
-                    this.Repaint();
-                }
-                
+                if (!string.IsNullOrEmpty(currentCategoryValue) && m_SpriteLibSelection.ContainsKey(currentCategoryValue))
+                    m_SpriteSelectorWidget.UpdateContents(m_SpriteLibSelection[currentCategoryValue].sprites);
+                else
+                    m_SpriteSelectorWidget.UpdateContents(Array.Empty<Sprite>());
+
+                Repaint();
+
                 m_PreviousCategoryValue = currentCategoryValue;
             }